change settings

+++ app/src/main/java/com/example/pomodorocompose/model/PomodoroSettings.kt
@@ -1,8 +1,8 @@
 package com.example.pomodorocompose.model
 
 data class PomodoroSettings(
-    val pomodoroDuration : Long,
-    val longRestDuration : Long,
-    val shortRestDuration : Long,
+    val pomodoroDuration : Int,
+    val longRestDuration : Int,
+    val shortRestDuration : Int,
     val pomodoroLoops : Int
 )
